Android app to calculate a distance from home
This app was written for Android-based devices (i.e. tablets and smartphones), and was written in XML (for the user interface) and Java (for the code behind). It uses GPS to calculate the distance between your current location and the location you set as your home point.

It will show the distance "as the crow flies", i.e. a straight line distance. The code indicates if you are within or outside the 2km from home, by means of text changing colour. Due to the Covid-19 lockdown that was introduced in Ireland on the 27th March 2020, you are supposed to stay within 2km of your home if you are just going for a walk or a cycle, except for specified essential tasks.